+++ libraries/FastLED/ci/ci/boards.py
@@ -11,6 +11,7 @@ ESP32_IDF_5_1_PIOARDUINO = "https://github.com/pioarduino/platform-espressif32/r
 
 # TODO: Upgrade toolkit to 5.3
 ESP32_IDF_5_3_PIOARDUINO = "https://github.com/pioarduino/platform-espressif32/releases/download/53.03.10/platform-espressif32.zip"
+ESP32_IDF_5_4_PIOARDUINO = "https://github.com/pioarduino/platform-espressif32/releases/download/54.03.20/platform-espressif32.zip"
 ESP32_IDF_5_1_PIOARDUINO_LATEST = (
     "https://github.com/pioarduino/platform-espressif32.git#develop"
 )
@@ -22,6 +23,8 @@ APOLLO3_2_2_0 = "https://github.com/nigelb/platform-apollo3blue"
 # Old fork that we were using
 # ESP32_IDF_5_1_PIOARDUINO = "https://github.com/zackees/platform-espressif32#Arduino/IDF5"
 
+# ALL will be auto populated in the Board constructor whenever a
+# board is defined.
 ALL: list["Board"] = []
 
 
@@ -156,7 +159,7 @@ ESP32_C2_DEVKITM_1 = Board(
     board_name="esp32c2",
     real_board_name="esp32-c2-devkitm-1",
     use_pio_run=True,
-    platform="https://github.com/Jason2866/platform-espressif32.git#Arduino/IDF5",
+    platform="https://github.com/pioarduino/platform-espressif32/releases/download/stable/platform-espressif32.zip",
     defines=["CONFIG_IDF_TARGET_ESP32C2=1"],
 )
 
@@ -175,14 +178,7 @@ ESP32_C6_DEVKITC_1 = Board(
 ESP32_S3_DEVKITC_1 = Board(
     board_name="esp32s3",
     real_board_name="seeed_xiao_esp32s3",  # Seeed Xiao ESP32-S3 has psram.
-    platform=ESP32_IDF_5_3_PIOARDUINO,
-    defines=[
-        "BOARD_HAS_PSRAM",
-    ],
-    build_flags=[  # Reserved for future use.
-        "-mfix-esp32-psram-cache-issue",
-        "-mfix-esp32-psram-cache-strategy=memw",
-    ],
+    platform=ESP32_IDF_5_4_PIOARDUINO,
     board_partitions="huge_app.csv",  # Reserved for future use.
 )
 
@@ -197,6 +193,13 @@ ESP32_H2_DEVKITM_1 = Board(
     board_name="esp32-h2-devkitm-1",
     platform_needs_install=True,  # Install platform package to get the boards
     platform=ESP32_IDF_5_3_PIOARDUINO,
+)
+
+ESP32_P4 = Board(
+    board_name="esp32p4",
+    real_board_name="esp32-p4-evboard",
+    platform_needs_install=True,  # Install platform package to get the boards
+    platform="https://github.com/pioarduino/platform-espressif32/releases/download/stable/platform-espressif32.zip",
 )
 
 ADA_FEATHER_NRF52840_SENSE = Board(