+++ libraries/FastLED/src/fl/tile2x2.h
@@ -5,6 +5,7 @@
 
 #include "fl/geometry.h"
 #include "fl/namespace.h"
+#include "fl/pair.h"
 #include "fl/slice.h"
 #include "fl/xymap.h"
 
@@ -31,6 +32,8 @@ class Tile2x2_u8 {
 
     void scale(uint8_t scale);
 
+    void setOrigin(int16_t x, int16_t y) { mOrigin = vec2<int16_t>(x, y); }
+
     uint8_t &operator()(int x, int y) { return at(x, y); }
     uint8_t &at(int x, int y) { return mTile[y][x]; }
     const uint8_t &at(int x, int y) const { return mTile[y][x]; }
@@ -40,33 +43,19 @@ class Tile2x2_u8 {
     uint8_t &lower_right() { return at(1, 0); }
     uint8_t &upper_right() { return at(1, 1); }
 
-    uint8_t maxValue() const {
-        uint8_t max = 0;
-        max = MAX(max, at(0, 0));
-        max = MAX(max, at(0, 1));
-        max = MAX(max, at(1, 0));
-        max = MAX(max, at(1, 1));
-        return max;
-    }
+    const uint8_t &lower_left() const { return at(0, 0); }
+    const uint8_t &upper_left() const { return at(0, 1); }
+    const uint8_t &lower_right() const { return at(1, 0); }
+    const uint8_t &upper_right() const { return at(1, 1); }
 
-    static Tile2x2_u8 Max(const Tile2x2_u8 &a, const Tile2x2_u8 &b) {
-        Tile2x2_u8 result;
-        for (int x = 0; x < 2; ++x) {
-            for (int y = 0; y < 2; ++y) {
-                result.at(x, y) = MAX(a.at(x, y), b.at(x, y));
-            }
-        }
-        return result;
-    }
+    uint8_t maxValue() const;
+
+    static Tile2x2_u8 MaxTile(const Tile2x2_u8 &a, const Tile2x2_u8 &b);
 
     vec2<int16_t> origin() const { return mOrigin; }
 
     /// bounds => [begin_x, end_x) (where end_x is exclusive)
-    rect<int16_t> bounds() const {
-        vec2<int16_t> min = mOrigin;
-        vec2<int16_t> max = mOrigin + vec2<int16_t>(2, 2);
-        return rect<int16_t>(min, max);
-    }
+    rect<int16_t> bounds() const;
 
     // Draws the subpixel tile to the led array.
     void draw(const CRGB &color, const XYMap &xymap, CRGB *out) const;
@@ -94,6 +83,26 @@ class Tile2x2_u8 {
     uint8_t mTile[2][2] = {};
     // Subpixels can be rendered outside the viewport so this must be signed.
     vec2<int16_t> mOrigin;
+};
+
+class Tile2x2_u8_wrap {
+    // This is a class that is like a Tile2x2_u8 but wraps around the edges.
+    // This is useful for cylinder mapping where the x-coordinate wraps around
+    // the width of the cylinder and the y-coordinate wraps around the height.
+    // This converts a tile2x2 to a wrapped x,y version.
+  public:
+    using Data = fl::pair<vec2i16, uint8_t>;  // absolute position, alpha
+
+    Tile2x2_u8_wrap() = default;
+    Tile2x2_u8_wrap(const Tile2x2_u8 &from, uint16_t width);
+    Tile2x2_u8_wrap(const Tile2x2_u8 &from, uint16_t width, uint16_t height);
+
+    // Returns the absolute position and the alpha.
+    Data &at(uint16_t x, uint16_t y);
+    const Data &at(uint16_t x, uint16_t y) const;
+
+  private:
+    Data tile[2][2] = {}; // zero filled.
 };
 
 } // namespace fl