+++ libraries/FastLED/release_notes.md
@@ -1,4 +1,18 @@
-FastLED 3.9.18 + 3.9.19
+FastLED 3.10.0
+==============
+
+  * Animartrix now out of beta.
+    * examples/Animartrix/Animartrix.ino
+  * ESP32
+    * Esp32P4 now officially supported.
+    * ESP32-S3 I2S driver is improved
+      * It will now auto error on known bad Esp32-Arduino Core versions.
+        * Arudino core 3.2.0 is now know to work.
+      * Documentation has been greatly simplified and unnecessary steps have been removed.
+
+
+
+FastLED 3.9.18 + 3.9.19 + 3.9.20
 ==============
 * Hotfixes for AVR platforms for 3.9.17